src/backend/requirements.txt
setuptools
wheel
meson
python-dateutil==2.8.2
Jetson.GPIO==2.0.21
awsiotsdk==1.11.9
marshmallow==3.19.0
pyyaml==5.4.1
grpcio==1.56.2
grpcio-tools==1.51.1
jsonschema==4.17.3
numpy==1.20.3
Pillow==10.3.0
pycairo==1.23.0
PyGObject==3.42.2
python-periphery==2.4.1
psutil==5.9.5
fastapi==0.109.2
uvicorn==0.23.2
pydantic==2.1.1
dlr==1.10.0
asgi-correlation-id==4.2.0
structlog==22.1.0
# https://pypi.org/project/Deprecated/
Deprecated==1.2.14
sqlalchemy==2.0.21
alembic==1.12.1
httpx==0.24.1